+++ src/Consts.ts
@@ -573,6 +573,29 @@ export default class Consts {
         '\t<UserJourneyBehaviors>\r\n' +
         '\t	<JourneyInsights TelemetryEngine="ApplicationInsights" InstrumentationKey="{instrumentationKey}" DeveloperMode="true" ClientEnabled="false" ServerEnabled="true" TelemetryVersion="1.0.0" />\r\n' +
         '\t</UserJourneyBehaviors>\r\n';
+
+        static CLAIM_stringCollection: string = '\t<ClaimType Id="{name}">\r\n' +
+        '\t	<DisplayName>{displayName}</DisplayName>\r\n' +
+        '\t	<DataType>stringCollection</DataType>\r\n' +
+        '\t	<AdminHelpText>Add help text here</AdminHelpText>\r\n' +
+        '\t	<UserHelpText>Add help text here</UserHelpText>\r\n' +
+        '\t</ClaimType>\r\n';
+
+
+        static CLAIM_Boolean: string = '\t<ClaimType Id="{name}">\r\n' +
+        '\t	<DisplayName>{displayName}</DisplayName>\r\n' +
+        '\t	<DataType>boolean</DataType>\r\n' +
+        '\t	<AdminHelpText>Add help text here</AdminHelpText>\r\n' +
+        '\t	<UserHelpText>Add help text here</UserHelpText>\r\n' +
+        '\t</ClaimType>\r\n';
+
+
+        static CLAIM_Integer: string = '\t<ClaimType Id="{name}">\r\n' +
+        '\t	<DisplayName>{displayName}</DisplayName>\r\n' +
+        '\t	<DataType>int</DataType>\r\n' +
+        '\t	<AdminHelpText>Add help text here</AdminHelpText>\r\n' +
+        '\t	<UserHelpText>Add help text here</UserHelpText>\r\n' +
+        '\t</ClaimType>\r\n';
 }