+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 15 May 2018
 - **Go to definition** - Always take precedence of editor open files, over unsaved version from file system (workspace folder)
-- **Add claim type** - New Paragraph claim type 
+- **Add claim type** - New Paragraph, String collection, Integer and Boolean claim types 
 
 ## 3 May 2018
 - **Go to definition** - If the element is not found in the selected file. Or if the selected element points to another file (the XML element is overwrite). The extension search the definition is all open files.